include symbol section if section contains strings
In issue #359, a lookup is performed on .LC1 when it shouldn't be.  This
is because the .LC1 symbol is being included by
kpatch_regenerate_special_section() but because the .rodata.str1.8
section is unchanged, the underlying section is not included and
kpatch_create_dynamic_rela_sections() treats it as an unchanged local
symbol and tries to look it up in vmlinux which fails.

This commit causes kpatch_regenerate_special_section() to include the
underlying string section.

Fixes #359

Signed-off-by: Seth Jennings <sjenning@redhat.com>