Handle modprobe style module name in kpatch command
Right now "kpatch list" will list "loaded patch modules" in modprobe
format.  A "kpatch load" with the name in this format will fail because
the modprobe format name does not match the module file name.
Additionally, because modprobe converts hyphens to the underscores and
underscores are allowed in the module file name, it is trivial to
transform a file name to a module name but non-trivial to
reverse the transformation.

This commit creates a symbolic link at install time with the module name
targeting the module file.  This establishes a mapping mechanism and
allows kpatch to trivially support using the module name in kpatch
commands.

Signed-off-by: Seth Jennings <sjenning@redhat.com>