create-diff-object: rewrite static local variable correlation
Rewrite the static local variable correlation logic.  It now goes
through all the static locals in the original object rather than the
patched object, ensuring that each symbol in the original object has a
twin.  It adds a new restriction that static local variables can't be
removed.

This fixes the following issues:

- Multiple static locals with the same name in the same function not
  allowed

- Two distinct static locals could be incorrectly correlated if they
  happen to have the same numbered suffix

Fixes: #514