kpatch-build: allow to specify the target kernel version
If the source tree of the kernel is used rather than its SRPM, the
version of the currently running kernel will be assumed as the version
of the target kernel (ARCHVERSION=$(uname -r)). This can be
inconvenient if the user actually prepares the binary patch for a
different kernel: the paths to some files used in the build contain
$ARCHVERSION, etc.

This patch adds -k|--kernel-version option to kpatch-build, so that it
becomes possible to specify the target kernel version explicitly.

Signed-off-by: Evgenii Shatokhin <eshatokhin@odin.com>