correct the correlation of the unlikely section
Sometimes, a function has two sections: .text.xxx and
.text.unlikely.xxx, the size of the unlikely one is zero.
Now the function changes from .text.xxx section to
.text.unlikely.xxx section, as follows.

old			new   (0: the section size is 0)
========================================================
.text.xxx
.text.unlikely.xxx(0)	.text.unlikely.xxx

So we just cross-correlate the sections, but must do the
uncorrelation logic of these sections before.

Additionly, the addralign of .text.unlikely.xxx in the new
object is 1, so just ignore this addralign check in this
situation.

Signed-off-by: Zhou Chengming <zhouchengming1@huawei.com>