kpatch: convert load_module() to use local variables
The kpatch script's functions don't use local variables, which makes it
bug prone and hard to understand.  It would be nice to convert the whole
file, but I'll just start with converting load_module() for now.

This also removes the unused 2nd argument to insmod, which is leftover
from commit 8e8de4718d8c ("kpatch: deprecate the replace command").