support dup file+symbol
This patch needs some explanation. When we want to resolve a dup
file+symbol symbol, we first find out our base object in lookup table,
then we can use this information to resolve ambiguous local symbols.

We use resolve_dup_file_symbol to find out the correct base object in
lookup table. If all local symbols of a file can be found in the
kelf_base, we think it's a matched file.

Fixes #604.

Signed-off-by: Zhou Chengming <zhouchengming1@huawei.com>