macros: add KPATCH_REVERT_FUNCTION macro
Add a KPATCH_REVERT_FUNCTION macro for forcing a function to be patched
with its original contents.  This is needed for the case where the patch
is a cumulative upgrade to a previous patch which modified the function,
and now the current patch needs to revert that change.

This is basically a copy/paste/modify job from KPATCH_IGNORE_FUNCTION
and kpatch_mark_ignored_functions_same().