kpatch: wait for livepatch transitions, poke stragglers
When loading a livepatch, wait for the patching transition to complete
within a reasonable timeframe, then poke any stalled tasks with a
signal.  If the transition is still taking too long, reverse the patch
and unload the livepatch.

When re-enabling a livepatch, do the same wait and signaling.  If the
expected time expires, disable the livepatch.

When unloading a livepatch, perform the wait/signaling, but only emit an
error message if the transition exceeds the time limit.

Signed-off-by: Joe Lawrence <joe.lawrence@redhat.com>